:sparkles: add user group cleanup on ISA revert to draft.

+++ api/src/services/information-sharing-agreements/destroy-groups-service.ts
@@ -1,6 +1,6 @@
 import { isNil } from "lodash"
 
-import { Group, InformationSharingAgreement, User } from "@/models"
+import db, { Group, InformationSharingAgreement, UserGroup, User } from "@/models"
 import BaseService from "@/services/base-service"
 
 export class DestroyGroupsService extends BaseService {
@@ -13,11 +13,25 @@ export class DestroyGroupsService extends BaseService {
 
   async perform(): Promise<void> {
     const { sharingGroupId, receivingGroupId } = this.informationSharingAgreement
-    const existingGroupIds = [sharingGroupId, receivingGroupId].filter((id) => !isNil(id))
-    await Group.destroy({
-      where: {
-        id: existingGroupIds,
-      },
+    if (isNil(sharingGroupId)) {
+      throw new Error("Sharing group ID is required")
+    }
+
+    if (isNil(receivingGroupId)) {
+      throw new Error("Receiving group ID is required")
+    }
+
+    return db.transaction(async () => {
+      await UserGroup.destroy({
+        where: {
+          groupId: [sharingGroupId, receivingGroupId],
+        },
+      })
+      await Group.destroy({
+        where: {
+          id: [sharingGroupId, receivingGroupId],
+        },
+      })
     })
   }
 }
